Patched some features and issues
- Added roles permission
- Added avatar upload
- Added accounts wizard
- Fixed LaTeX rendering issue
- Fixed printing rendering
- Removed unique key of 'name' in 'courses' table

+++ courses.js
@@ -481,7 +481,10 @@ const Courses =
             try
             {
                 const path = "./src/banners/" + id;
-                FileIO.unlinkSync(path);
+                
+                if (FileIO.existsSync(path))
+                    FileIO.unlinkSync(path);
+
                 return true;
             }
             catch(error)
